led_driver: Change LED driver timer frequency to 12 KHz
On STM32 we used 12 KHz LED driver timer frequency:
  parent_freq = 48 MHz
  prescaler = 20
  period = 200
  freq = parent_freq / prescaler / period = 12 KHz

In order to be completely compatible, we need to use the same frequency
on GD32. We can achieve this by using period 300 and prescaler 20. This
way only the period is different from STM32 version.

Do the same in bootloader's LED driver.

Signed-off-by: Marek Behún <marek.behun@nic.cz>