Makefile: Use optimize for size
We need to use -Os in order to have a chance to put all code into the
first half of GD32's flash memory, which is needed because the second
half is slow (there are long delays when fetching instructions).

Since our code resides at a specific address, we can also add -fno-pie,
which saves some more space.

Signed-off-by: Marek Behún <marek.behun@nic.cz>