led_driver: Fix nSS latching
Now that whole code is in the first half of the flash memory on GD32 and
the code executes without delay when fetching instructions, together
with enabled optimizations, it seems that the nSS pin latching is not
working correctly and this causes unpredictable colors and even LED
addressing.

This is because in the code
  LATCH_HIGH;
  __NOP();
  __NOP();
  __NOP();
  LATCH_LOW;
there is too little delay (3 NOPs) between latch high and latch low.

Rework this code such that:
- on initialization, nSS is put high,
- only when new RGB triplet is to be sent is nSS put low, but before
  frame data computation (the computation takes some instructions and
  thus generates a natural delay),
- after whole RGB triplet is sent, nSS is put high.

Do the same in bootloader's LED driver.

Signed-off-by: Marek Behún <marek.behun@nic.cz>