slave_i2c_device: Simplify reading I2C status register
The i2c_interrupt_flag_get() has a rather complex code for what it does,
and the reason is simply bad design of macro constants in the GD32
library.

Read the I2C status 0 register directly in the interrupt handler, which
makes the handler faster and less complicated (even for reading the
code!).

Simplify also clearing of the ADDSEND bit.

Signed-off-by: Marek Behún <marek.behun@nic.cz>