slave_i2c_device: Prepare I2C handler state machine
On GD32 sometimes when the I2C interrupt handler is invoked, multiple
interrupts can be set.

If these happen to be STPDET and ADDSEND because a transaction was
completed (STPDET) and new trasaction immediately started (ADDSEND), we
need to somehow enforce handling of STPDET first.

Prepare a handler state machine for this.

Signed-off-by: Marek Behún <marek.behun@nic.cz>