app, bootloader: Fix interrupt priority group
For all interrupts we are using pre-emption priority 0 and
sub-priorities 1-5, but GD32's implementation of nvic_enable_irq()
allows for only 4 bits of priority, and the default setting is 2 bits
for pre-emption priority and 2 bits for subpriority.

This means that upper bits of the subpriorities we are using are throws
away, and so our prioritization

  handler     prio                       handler     prio
  =======     ====                       =======     ====
  i2c         1                          leds        0
  delay       2       ==============>    i2c         1
  debounce    3       gets changed to    led effect  1
  leds        4                          power       1
  led effect  5                          delay       2
  power       5                          debounce    3

Fix this by allowing one more bit for subpriority.

Signed-off-by: Marek Behún <marek.behun@nic.cz>