app, boot: Store application's crc32 and check in bootloader
GD32 offers HW CRC32 circuit. When compiling application binary, store
its size and crc32 checksum at offsets 0x110 and 0x114 of the binary
(immediately after ISR vectors).

Then in bootloader if the crc32 checksum does not check out, do not boot
the application and stay in bootloader, but power the device on.

Signed-off-by: Marek Behún <marek.behun@nic.cz>