bootloader: Fix reset button / reboot after flashing with mcutool
The reset button does not currenlty work when in bootloader.

Also after flashing MCU firmware with mcutool and rebooting, the reboot
does not work.

This is because in bootloader's RESET_MANAGER code, we need to read
SYSRES_OUT_PIN's input, not output.

In STM32's version we are reading input.

I guess this was a mistake when converting from STM32 to GD32.

Signed-off-by: Marek Behún <marek.behun@nic.cz>