slave_i2c_device: Rename SFP_DIS_MASK to RESERVED_MASK
The SFP_DIS bit from general control command was historically used to
control the TX_DISABLE pin of SFP cage on a prototype board. The code
to handle this was then removed when the pin was connected to GPIO
expander where it is controllable from the SOC.

The documentation says "don't care" about this bit.

Rename the constant.

Signed-off-by: Marek Behún <marek.behun@nic.cz>