Add ability to automatically switch between PHY/SFP WAN modes
In order to be compatible with older revisions of the board we need by
default to automatically switch between PHY/SFP WAN modes according to
the value of SFP_DET pin.

Add PHY_SFP_AUTO bit to exnteded control (with default value of 1). When
set, automatic switching occurs. When unset, the value written to
PHY_SFP bit of extended control is respected.

Signed-off-by: Marek Behún <marek.behun@nic.cz>