Don't yet change RGB LED on WAN LED1 pin change
Commit ac6ed7c2943d ("add WAN LED1 pin functionality") added WAN LED1
pin support since it is connected on board revision 32 and later.

It made it so that the RGB LED turns on if WAN LED1 pin is active. We do
not want that, since the default setting for WAN LED1 pin isn't
configured correctly yet, and also because we want to add pin mapping
functionality, which isn't supported yet.

So do not change RGB LED on WAN LED1 pin activity for now.

Signed-off-by: Marek Behún <marek.behun@nic.cz>