drivers/led_driver: Use more color levels when gamma correcting
Since we have a nicely optimized frame computing code (~163 cycles on
Cortex-M0, ~138 cycles on Cortex-M3), we can call it more often.

This is useful when doing gamma correction of colors: when using only
256 color levels, with gamma correction the low value (0 - 10) grow very
slowly, meaning that changing color from #020300 to #050a00 does not
change the intensity of the LEDs at all.

Let's increase color levels 2.5 times on STM32 and 4 times on GD32
(since it is faster). Do this only when gamma correction is enabled.

Signed-off-by: Marek Behún <marek.behun@nic.cz>