bootloader: Blink red color if booting application failed
We blink green color when in bootloader, to indicate to the user that
MCU is waiting for application flashing.

This is okay for when the user entered bootloader from application via
general control command.

But if the application was flashed incorrectly and cannot be booted, we
should instead blink different color to indicate this error. Let's blink
with red.

Signed-off-by: Marek Behún <marek.behun@nic.cz>