power_control: Fix reset selector speed
Reset selector is noticeably slower on STM32 due to usage of msleep(1).
With value 1 it calls mdelay(), and we didn't account for the fact that
interrupts are not disabled at this stage, and we spend a lot of time in
LED interrupt handler. This means that mdelay() delays for 2x-3x more on
STM32.

Make away with the msleep(1) call. Instead put a cycle of some nops
in-between reading the SYSRES GPIO, and detect the timeout for
increasing selector by comparing jiffies.

Signed-off-by: Marek Behún <marek.behun@nic.cz>