drivers/led_driver: Fix LED_TIMER_PERIOD value
Commit d584227619bd ("drivers/led_driver: Use more color levels when
gamma correcting") introduced wrong value of 5 for LED_TIMER_PERIOD,
resulting in the led_driver_send_frame() interrupt to be called too fast
when gamma correction is disabled. On STM32, for example, the interrupt
is called every 250 cycles instead of every 1000, as described by the
comments added in the same commit.

Fix LED_TIMER_PERIOD definition.

Fixes: d584227619bd ("drivers/led_driver: Use more color levels when gamma correcting")
Signed-off-by: Marek Behún <marek.behun@nic.cz>