timer: Change semantics of timer_init() and timer_set_period/freq()
Change the semantics of the timer_init(type=timer_interrupt) function.

Currently the actual frequency of the timer interrupt is freq / period
(freq and period are parameters to the timer_init function). This is due
to how on both STM32 and GD32 the timers have a 16-bit prescaler and
a 16-bit counter auto reload register (we put SYS_CORE_FREQ / freq into
the prescaler register and period into the counter auto reload register).

We change the semantics so that the period parameter isn't used (must be
zero) and the freq parameter represents the actual frequency with which
the timer will produce the interrupt.

We remove the timer_set_period() function and change timer_set_freq()
accordingly.

Signed-off-by: Marek Behún <marek.behun@nic.cz>