drivers/led_driver: Fix reset effect when interrupted during progress
There is an issue wherein when reset effect is in progress and it is
interrupted by the reset button being pressed, the new reset effect will
continue where the last reset effect was interrupted, instead of
starting from the beginning.

This is due to the static variables led and effect_timeout_cnt not being
initalized in EFFECT_INIT event.

Signed-off-by: Marek Behún <marek.behun@nic.cz>