drivers/led_driver: Fix reset effect toggling
Refactor enabling/disabling of the reset effect. Drop the
effect_reset_finished global variable, and instead only use the value of
reset_effect_state variable to determine whether reset effect is in
progress (we add another value, EFFECT_DISABLED, to the enumerator, to
indicate that the effect is not in progress).

Fix toggling of reset effect by changing the value of reset_effect_state
when calling led_driver_reset_effect.

Signed-off-by: Marek Behún <marek.behun@nic.cz>