drivers/led_driver: Support overwriting PWM brightness
Add the ability to temporarily overwrite PWM brightness and use it in
power_control_first_startup().

Currently in power_control_first_startup() we first save the previous
value of PWM brightness, then play with it, and then return in back to
the cached value. But if in the meantime user changes the value via I2C
slave interface, the pattern we are making will be disrupt, and
afterwards we will disregard the new value the user requested in favor
to the old one.

By adding a special variable indicating whether the brightness is
currently overwritten we avoid this issue.

Signed-off-by: Marek Behún <marek.behun@nic.cz>