Implement memory protection on MKL81 platform
The MKL81 microcontroller, which has a hardware cryptographic
acceleration unit and true random number generator, was chosen to be
used on new Turris Omnia boards in order to replace the functionality
provided by the ATSHA204A cryptochip. We want it to implement:
- storage of MAC address and serial number,
- storage of secret cryptographic keys,
- authentication of the board (which is to be implemented by asymmetric
  cryptographic signatures),
- true random number generation.

The needed cryptographic routines and the I2C commands that expose the
functionality to Turris Omnia's SoC are yet to be implemented.

But in preparation of this, we now implement and start using memory
protection in MKL81 firmware, in order to prevent the leaking of secret
cryptographic keys.

We implement:
- usage of unprivileged mode with separate stack,
- memory protection ensuring that executable sections are not writable
  in unprivileged mode, that unprivileged mode cannot access privileged
  data,
- peripherals configuration ensuring that unprivileged mode can only use
  allowed peripherals (GPIOs, CRC unit, timers, SPI for LEDs, UART, I2C
  and RFSYS),
- supervisor calls to allow unprivileged mode to call privileged system
  calls when it needs to do a privileged operation,
- ability to securely verify the signature of new firmware in the
  firmware flashing handlers (once cryptographic routines are
  implemented).

Since the main place of a possible attack are potential bugs in the I2C
command handlers, we put the I2C driver and I2C command parsing code
into unprivileged mode.

Signed-off-by: Marek Behún <marek.behun@nic.cz>