bootloader: Initialize I2C interface only when needed
Do not initialize the I2C slave interface unconditionally. Only do so
when needed, which is either when the bootloader is powering up the
device (because application image is corrupted), or when the application
requested to stay in bootloader to be able to flash new application
image.

We can move the i2c_iface_init() + i2c_slave_init() sequence to
to when interrupts are enabled because since commit 48bcb221197f
("Disable interrupts and clear pending interrupts in startup") we ensure
that during this sequence the I2C interrupt does not fire.

Signed-off-by: Marek Behún <marek.behun@nic.cz>